Enhance documentation and examples; implement session-based nonce handling
- Update README with detailed architecture and wire format specifications.
- Add Python client example to demonstrate interaction with the C client library.
- Implement session-based nonce handling in core logic to prevent replay attacks.
- Modify server to persist entries and expose an endpoint for retrieving them.
- Update Makefile to include shared library build.
- Create setup script for generating device fixtures.
- Refactor tests to accommodate new nonce handling and session management.

+++ core/src/types.rs
@@ -13,7 +13,7 @@ pub struct VerifyingKey {
 /// Log entry as defined in the wire format.
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub struct LogEntry {
-    pub version: f32,
+    pub version: u8,
     pub entry_hash: String,
     pub device_id: String,
     pub timestamp: i64,
